more zeta_zero tests, docs, citations, and comments
