+++ acb_dirichlet.h
@@ -168,6 +168,12 @@ void acb_dirichlet_hardy_z_series(acb_poly_t res, const acb_poly_t s, const diri
 void acb_dirichlet_gram_point(arb_t res, const fmpz_t n, const dirichlet_group_t G, const dirichlet_char_t chi, slong prec);
 void acb_dirichlet_backlund_s_bound(mag_t res, const arb_t t);
 ulong acb_dirichlet_turing_method_bound(const fmpz_t p);
+int _acb_dirichlet_definite_hardy_z(arb_t res, const arf_t t, slong *pprec);
+void _acb_dirichlet_isolate_gram_hardy_z_zero(arf_t a, arf_t b, const fmpz_t n);
+void _acb_dirichlet_isolate_rosser_hardy_z_zero(arf_t a, arf_t b, const fmpz_t n);
+void _acb_dirichlet_isolate_turing_hardy_z_zero(arf_t a, arf_t b, const fmpz_t n);
+void acb_dirichlet_isolate_hardy_z_zero(arf_t a, arf_t b, const fmpz_t n);
+void acb_dirichlet_hardy_z_zero(arb_t res, const fmpz_t n, slong prec);
 void acb_dirichlet_zeta_zero(acb_t res, const fmpz_t n, slong prec);
 
 /* Discrete Fourier Transform */
